Move `assign_label_colors_by_groups` from `qc.py` to `qc2.py`
